+++ rewrite/changelog.md
@@ -20,6 +20,7 @@ The subsections are formatted as they need when publishing a release to the Play
 ## Unreleased
 
  * Increased log file max size to 32k
+ * Added MoStreamingService.IsAwakeLogThread
 
 ## [0.9.21] - 2019-8-17 12:25